BREAKING port/include-path change: add SPI transport and common type API.
An SPI API is added in the porting layer and implemented on the Zephyr, ESP-IDF/[Arduino], NRF5SDK and STM32F4 platforms.  It is tested for use with u-blox GNSS chips.

IF YOU ARE NOT USING the provided ubxlib.cmake or ubxlib.make files, you will need to add a NEW INCLUDE PATH, common/type/api, to your build.  This include directory is for type definitions that straddle layers of ubxlib.  It currently contains common_spi.h, which is needed by both u_port_spi.h and u_device.h.

IF YOU HAVE YOUR OWN PORTING LAYER you can return U_ERROR_COMMON_NOT_SUPPORTED for all of the functions in the port SPI API without any adverse effect, unless of course you happen to want to use it to talk with a GNSS chip.