SPI support for GNSS.
*** WARNINING *** WARNING *** WARNING *** WARNING ***

This is a preview branch that has NOT been heavily tested and the SPI-related contents of which (i.e. the three commits before this) remain SUBJECT TO CHANGE.  It may be of use if you really need to connect a u-blox GNSS chip to your MCU via SPI and you can't wait for a properly tested version of this change to be available on master (not before early 2023 I'm afraid).

*** WARNINING *** WARNING *** WARNING *** WARNING ***